dogfood dotnet_test
