--- dotnet/tools/builder/Builder.sln
@@ -1,28 +1,0 @@
-﻿# todo(#12) delete this file
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "builder", "builder.csproj", "{F9FDB2CC-41E9-4C73-BF95-91B44FCB96C4}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "runfiles", "..\runfiles\runfiles.csproj", "{ED862780-5AE5-4F3E-9DB3-08C6B9572DA7}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BuilderTests", "..\..\..\tests\dotnet\tools\builder\BuilderTests.csproj", "{28583800-39CB-4643-9FF0-071F5776D1DB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F9FDB2CC-41E9-4C73-BF95-91B44FCB96C4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F9FDB2CC-41E9-4C73-BF95-91B44FCB96C4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F9FDB2CC-41E9-4C73-BF95-91B44FCB96C4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F9FDB2CC-41E9-4C73-BF95-91B44FCB96C4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ED862780-5AE5-4F3E-9DB3-08C6B9572DA7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ED862780-5AE5-4F3E-9DB3-08C6B9572DA7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ED862780-5AE5-4F3E-9DB3-08C6B9572DA7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ED862780-5AE5-4F3E-9DB3-08C6B9572DA7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{28583800-39CB-4643-9FF0-071F5776D1DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{28583800-39CB-4643-9FF0-071F5776D1DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{28583800-39CB-4643-9FF0-071F5776D1DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{28583800-39CB-4643-9FF0-071F5776D1DB}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
