--- dotnet/tools/builder/builder.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <!-- todo(#12) delete this file -->
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>MyRulesDotnet.Tools.Builder</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\runfiles\runfiles.csproj" />
-  </ItemGroup>
-
-</Project>
