+++ dotnet/tools/runfiles/BazelEnvironment.cs
@@ -7,8 +7,7 @@ namespace MyRulesDotnet.Tools
     {
         public static string GetTestTmpDir()
         {
-            // todo(#12) remove Path.GetTempPath()
-            var value = Environment.GetEnvironmentVariable("TEST_TMPDIR") ?? Path.GetTempPath();
+            var value = Environment.GetEnvironmentVariable("TEST_TMPDIR");
             if (string.IsNullOrEmpty(value)) throw new IOException("$TEST_TMPDIR is empty or undefined");
             return value;
         }
