--- dotnet/tools/runfiles/Runfiles.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "runfiles", "runfiles.csproj", "{D8FEE1E0-433F-4AC7-844C-589231A6FC97}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RunfilesTests", "..\..\..\tests\dotnet\tools\runfiles\RunfilesTests.csproj", "{8427891A-FD2E-491C-A59F-A2BF05D97FDA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D8FEE1E0-433F-4AC7-844C-589231A6FC97}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D8FEE1E0-433F-4AC7-844C-589231A6FC97}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D8FEE1E0-433F-4AC7-844C-589231A6FC97}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D8FEE1E0-433F-4AC7-844C-589231A6FC97}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8427891A-FD2E-491C-A59F-A2BF05D97FDA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8427891A-FD2E-491C-A59F-A2BF05D97FDA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8427891A-FD2E-491C-A59F-A2BF05D97FDA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8427891A-FD2E-491C-A59F-A2BF05D97FDA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
