--- tests/dotnet/tools/runfiles/DotnetCat.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.IO;
-
-namespace MyRulesDotnet.Tools.RunfilesTests
-{
-    public static class DotnetCat
-    {
-        static void Main(string[] args)
-        {
-            var r = Runfiles.Create();
-            var contents = File.ReadAllText(r.Rlocation("my_rules_dotnet/tests/dotnet/tools/runfiles/foo.txt"));
-
-            Console.Write(contents);
-        }
-    }
-}
