--- tests/dotnet/tools/runfiles/RunfilesTests.csproj
@@ -1,26 +1,0 @@
-<!-- todo(#12) remove this file -->
-<Project Sdk="Microsoft.NET.Sdk">
-  
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>MyRulesDotnet.Tools.RunfilesTests</RootNamespace>
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="FluentAssertions" Version="5.10.3" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.5.0" />
-    <PackageReference Include="xunit" Version="2.4.0" />
-    <PackageReference Include="xunit.runner.visualstudio" Version="2.4.0" />
-    <PackageReference Include="coverlet.collector" Version="1.2.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\..\dotnet\tools\runfiles\runfiles.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Remove="DotnetCat.cs" />
-  </ItemGroup>
-
-</Project>
