+++ tests/dotnet/tools/runfiles/integration/dotnet_cat_test.py
